Working on CSS. Creating basic structure.
I spent a lot of time trying to automate the Less compiling. We probably
don't even need Less to be honest, but since I've put in the work and it
might be useful later, I'm keeping it. Less compiling is automated now
via less-watch-compiler version 1.8.0. It does some weird things, but
the latest version has a bug that makes it unhelpful for me. I'll wait
and see if that gets fixed soon.